move courts pages helper methods into their own module
